Restructure formatters
