+++ pyproject.toml
@@ -39,9 +39,6 @@ dependencies = [
     "aws-sam-translator>=1.69.0",
     "jsonpatch",
     "networkx>=2.4,<4",
-    "junit-xml~=1.9",
-    "jschema_to_python~=1.2.3",
-    "sarif-om~=1.0.4",
     "sympy>=1.0.0",
     "regex",
     "typing_extensions",
@@ -50,6 +47,13 @@ dependencies = [
 [project.optional-dependencies]
 graph = [
     "pydot"
+]
+junit = [
+    "junit-xml~=1.9"
+]
+sarif = [
+    "jschema_to_python~=1.2.3",
+    "sarif-om~=1.0.4",
 ]
 
 [project.scripts]
