+++ src/cfnlint/formatters/BaseFormatter.py
@@ -1,0 +1,23 @@
+"""
+Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+SPDX-License-Identifier: MIT-0
+"""
+
+
+class BaseFormatter:
+    """Base Formatter class"""
+
+    def _format(self, match):
+        """Format the specific match"""
+
+    def print_matches(self, matches, rules, config):
+        """Output all the matches"""
+        if not matches:
+            return None
+
+        # Output each match on a separate line by default
+        output = []
+        for match in matches:
+            output.append(self._format(match))
+
+        return "\n".join(output)
