+++ src/cfnlint/formatters/Formatter.py
@@ -1,0 +1,21 @@
+"""
+Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+SPDX-License-Identifier: MIT-0
+"""
+
+from cfnlint.formatters.BaseFormatter import BaseFormatter
+
+
+class Formatter(BaseFormatter):
+    """Generic Formatter"""
+
+    def _format(self, match):
+        """Format output"""
+        formatstr = "{0} {1}\n{2}:{3}:{4}\n"
+        return formatstr.format(
+            match.rule.id,
+            match.message,
+            match.filename,
+            match.linenumber,
+            match.columnnumber,
+        )
